feat(#80): PR 3 — spawn module extraction + role registry
Move the spawn machinery (~310 LOC) out of `lib/lore_cli/hooks.py`
into a new `lib/lore_cli/spawn.py`. The new module owns the cooldown
stamp primitives, log + meta-sidecar rotation, the runaway-process
gate, the core `_spawn_detached`, and the Curator-C scheduling helpers
(`_curator_c_email`, `_curator_c_jitter_seconds`, `_iso_week_monday_utc`).

Adds the `SpawnRole` dataclass + `SPAWN_ROLES` registry and a single
public `spawn(role, lore_root, *, cooldown_s=None, **extra) -> bool`
entry point per the PRD. Each role row encodes `name`, `argv_builder`,
`default_cooldown_s`, and `migrate_stamp` — adding a new background
process is one row. The four legacy `_spawn_detached_curator_{a,b,c}`
and `_spawn_detached_transcript_sync` wrappers are kept as one-liners
that delegate to `spawn()`, preserving every call site and every test
patch point on `lore_cli.hooks._spawn_detached_curator_*`.

`hooks.py` now re-exports the spawn names so internal heartbeat code
and external test patches keep resolving in the `lore_cli.hooks`
namespace; spawn-core machinery stays untouched per the PRD no-touch
list (this is a move + registry, not a behaviour change).

Test surface:
- `tests/test_proc_log_capture.py` and `tests/test_spawn_runaway_detector.py`
  needed their patch targets repointed from `lore_cli.hooks.<helper>`
  to `lore_cli.spawn.<helper>` because those tests patch helpers that
  `_spawn_detached` looks up in its own module globals; the patches in
  the hooks namespace no longer reach the spawn-side bindings.
- All other spawn / hook tests (38 of them) work unchanged because they
  patch the public wrappers that hooks.py still re-exports.

Net delta: hooks.py shrinks by 309 LOC; spawn.py adds 444 (the cost of
the dataclass + registry + module docstring + section headers, paid
once). Net +142 LOC across two files but the cohesion gain in hooks.py
is the visible win — `wc -l hooks.py` drops 3124 → 2815.

Validation: full pytest suite is **2634 passed, 7 skipped** (two
pre-existing failures still unrelated: missing `openai` optional dep,
`test_resume` fixture date drift). `lore hook --help` renders the same
subcommand surface; `SPAWN_ROLES` exposes `{a, b, c, transcripts}` with
cooldowns `{60, 300, 3600, 300}` and `migrate_stamp` flags
`{True, True, False, False}` — byte-stable with the old wrappers.

Refs #80
